Move log files into a its own folder
